refactor(model): update some models
models: "RoomStatus.java" "Amenity.java" "Reservation.java" "Room.java"
