+++ src/backend/Roomly/src/main/java/org/example/roomly/model/RoomStatus.java
@@ -1,0 +1,5 @@
+package org.example.roomly.model;
+
+public enum RoomStatus {
+    Available, Unavailable
+}
